+++ src/dist_s1/__main__.py
@@ -113,6 +113,28 @@ def common_options(func: Callable) -> Callable:
         required=False,
         help='S3 bucket prefix to upload the final products to.',
     )
+    @click.option(
+        '--device',
+        type=click.Choice(['cpu', 'cuda', 'mps', 'none']),
+        required=False,
+        default=None,
+        help='Device to use for transformer model inference of normal parameters.',
+    )
+    @click.option(
+        '--batch_size_for_despeckling',
+        type=int,
+        default=25,
+        required=False,
+        help='Batch size for despeckling the bursts; i.e. how many arrays are loaded into CPU memory at once.',
+    )
+    @click.option(
+        '--n_workers_for_norm_param_estimation',
+        type=int,
+        default=1,
+        required=False,
+        help='Number of CPUs to use for normal parameter estimation; error willbe thrown if GPU is available and not'
+        ' or set to something other than CPU.',
+    )
     @functools.wraps(func)
     def wrapper(*args: P.args, **kwargs: P.kwargs) -> R:
         return func(*args, **kwargs)
@@ -149,6 +171,9 @@ def run_sas_prep(
     bucket: str | None,
     bucket_prefix: str,
     n_workers_for_despeckling: int,
+    batch_size_for_despeckling: int,
+    n_workers_for_norm_param_estimation: int,
+    device: str | None,
 ) -> None:
     """Run SAS prep workflow."""
     run_config = run_dist_s1_sas_prep_workflow(
@@ -169,6 +194,9 @@ def run_sas_prep(
         bucket=bucket,
         bucket_prefix=bucket_prefix,
         n_workers_for_despeckling=n_workers_for_despeckling,
+        batch_size_for_despeckling=batch_size_for_despeckling,
+        n_workers_for_norm_param_estimation=n_workers_for_norm_param_estimation,
+        device=device,
     )
     run_config.to_yaml(runconfig_path)
 
@@ -202,6 +230,9 @@ def run(
     bucket: str | None,
     bucket_prefix: str,
     n_workers_for_despeckling: int,
+    batch_size_for_despeckling: int,
+    n_workers_for_norm_param_estimation: int,
+    device: str | None,
 ) -> str:
     """Localize data and run dist_s1_workflow."""
     # Localize data
@@ -223,6 +254,9 @@ def run(
         bucket=bucket,
         bucket_prefix=bucket_prefix,
         n_workers_for_despeckling=n_workers_for_despeckling,
+        batch_size_for_despeckling=batch_size_for_despeckling,
+        n_workers_for_norm_param_estimation=n_workers_for_norm_param_estimation,
+        device=device,
     )
     return run_config